requirements.txt
aioredis==2.0.1
aiosmtplib==3.0.2
alembic==1.16.5
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Deprecated==1.3.1
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
fastapi-mail==1.4.2
greenlet==3.2.5
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
limits==4.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==24.2
passlib==1.7.4
pillow==11.3.0
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pyinstrument==5.1.2
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
redis==7.0.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
sqlmodel==0.0.34
starlette==0.49.3
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uvicorn==0.39.0
vine==5.1.0
wcwidth==0.6.0
wrapt==2.1.2
